% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/57_sysreqs_recipies.R
\name{build_bash_script}
\alias{build_bash_script}
\title{Creates a bash script to update the system to satisfy project requirements.}
\usage{
build_bash_script(recipe, plat_desc)
}
\arguments{
\item{recipe}{object of type sysreqs_script_recipe}

\item{plat_desc}{named list content:
\describe{
  \item{name}{One of Windows, MacOS, RedHat, Debian (type: character)}
  \item{distrib}{Distribution e.g. for Debian: Debian, Ubuntu; for RedHat: CentOS, RedHat, Fedora (type: character(1))}
  \item{release}{Distribution release e.g. for Debian: squeeze, wheezy, jessie (type: character(1))}
  \item{sysreq_type}{One of Windows, Pkg, RPM, DEB (type: character(1))}
  \item{build}{True if build environment is required (type: logical(1))}
}}
}
\description{
Creates a bash script to update the system to satisfy project requirements.
}
