% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40_ci_adapter.R
\name{ci_adapter_get_version}
\alias{ci_adapter_get_version}
\title{Retrieves current CI build number.}
\usage{
ci_adapter_get_version(ci_adapter)
}
\arguments{
\item{ci_adapter}{ci adapter object}
}
\value{
build number reported by CI. (type: character).
}
\description{
Retrieves current CI build number.
}
\examples{
# create you own CI adapter
ci_adapter_create_own <- function() {
  result <- ci_adapter_create_base("Own")
  class(result) <- c("ci_adapter_own", class(result))
  return(result)
}

#' @export
ci_adapter_get_version.ci_adapter_own <- function(ci_adapter) {
  # ... detect if build triggered by CI is currently running ...
  return("0.0")
}

}
\seealso{
Other in extending RSuite with CI adapter: \code{\link{ci_adapter_create_base}},
  \code{\link{ci_adapter_is_building}}
}
