% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40_ci_adapter.R
\name{ci_adapter_is_building}
\alias{ci_adapter_is_building}
\title{Detects if build process triggered by CI is currently running.}
\usage{
ci_adapter_is_building(ci_adapter)
}
\arguments{
\item{ci_adapter}{ci adapter object}
}
\value{
TRUE if build triggered by CI is currently running.
}
\description{
Detects if build process triggered by CI is currently running.
}
\examples{
# create you own CI adapter
ci_adapter_create_own <- function() {
  result <- ci_adapter_create_base("Own")
  class(result) <- c("ci_adapter_own", class(result))
  return(result)
}

#' @export
ci_adapter_is_building.ci_adapter_own <- function(ci_adapter) {
  # ... check ...
  return(TRUE)
}

}
\seealso{
Other in extending RSuite with CI adapter: \code{\link{ci_adapter_create_base}},
  \code{\link{ci_adapter_get_version}}
}
\concept{in extending RSuite with CI adapter}
