% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_install_prj_deps.R
\name{clean_prj_deps}
\alias{clean_prj_deps}
\title{Cleans installed packages that are not required from project local environment.}
\usage{
clean_prj_deps(params)
}
\arguments{
\item{params}{project parameters(type: rsuite_project_params)}
}
\description{
Cleans installed packages that are not required from project local environment.
}
\keyword{internal}
