% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/60_versions.R
\name{get_version_numbers}
\alias{get_version_numbers}
\title{Retrieves version numbers from the input version string e.g. 1.2.0
returns  c(1, 2, 0)}
\usage{
get_version_numbers(vers)
}
\arguments{
\item{vers}{list of versions which can contain blocks of digits separeded with a dot
or dash character (type: character).}
}
\value{
list of versions digit vectors (type: list)
}
\description{
Retrieves version numbers from the input version string e.g. 1.2.0
returns  c(1, 2, 0)
}
