% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_api_inst.R
\name{inst_wrap_zip}
\alias{inst_wrap_zip}
\title{Wraps deployment zip into bash installer script.}
\usage{
inst_wrap_zip(zip_fpath)
}
\arguments{
\item{zip_fpath}{path to zip package to wrap. It must exist. (type: character)}
}
\value{
path to created bash installer script (invisible)
}
\description{
Wraps deployment zip into bash installer script.
}
\details{
Bash installer is just script containing also binary data at the end of file.
It has also some logic to install the package into more intelligent way than just
unzip.

Shell script is created at the same location as zip package passed and will have
the same name with sh extension.
}
\examples{
\donttest{
  inst_wrap_zip("myproj_0.1-1.zip") # creates myproj_0.1-1.sh
}
}
\concept{in installer}
