% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/57_sysreqs_recipies.R
\name{perform}
\alias{perform}
\title{Performs(runs) all recipes from the sysreqs_recipe object.}
\usage{
perform(recipe)
}
\arguments{
\item{recipe}{sysreqs_recipe object (type: sysreqs_recipe)}
}
\description{
Performs(runs) all recipes from the sysreqs_recipe object.
}
