% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_install_prj_deps.R
\name{pkg_inst_order}
\alias{pkg_inst_order}
\title{Detects order of installation of packages as permutation of input vector}
\usage{
pkg_inst_order(pkgs, db)
}
\arguments{
\item{pkgs}{vector of package names to permute}

\item{db}{data frame of available packages as returned by available.packages}
}
\value{
numeric vector which is defining permutation of packages ta align
them in order of installation (less dependent to more dependent).
}
\description{
Detects order of installation of packages as permutation of input vector
}
\keyword{internal}
