% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30_rc_adapter.R
\name{rc_adapter_remove_admins}
\alias{rc_adapter_remove_admins}
\title{Remove all RC related administrative entries from folder tree at dir.}
\usage{
rc_adapter_remove_admins(rc_adapter, dir)
}
\arguments{
\item{rc_adapter}{rc adapter object}

\item{dir}{path to directory to remove administrators from. Folder must
exist (type: character)}
}
\description{
This is required for cleaning temporary folder during collecting entries to
put into project zip package.
}
\examples{
# create you own RC adapter
rc_adapter_create_own <- function() {
  result <- rc_adapter_create_base("Own")
  class(result) <- c("rc_adapter_own", class(result))
  return(result)
}

#' @export
rc_adapter_remove_admins.rc_adapter_own <- function(rc_adapter, dir) {
  # ... unlink RC administrative folders from dir (like .svn or .git) ...
}

}
\seealso{
Other in extending RSuite with RC adapter: \code{\link{rc_adapter_create_base}},
  \code{\link{rc_adapter_get_version}},
  \code{\link{rc_adapter_is_under_control}},
  \code{\link{rc_adapter_pkg_struct_add}},
  \code{\link{rc_adapter_prj_struct_add}}
}
