% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_api_rsuite.R
\name{rsuite_check_version}
\alias{rsuite_check_version}
\title{Checks if newer version of RSuite is available.}
\usage{
rsuite_check_version()
}
\value{
NULL if newer version is not available or newest available version number.
}
\description{
Checks if newer version of RSuite is available.
}
\examples{
\donttest{
  # print latest version available or NULL if latest is currently installed
  rsuite_check_version()
}

}
\seealso{
Other miscellaneous: \code{\link{rsuite_getLogger}},
  \code{\link{rsuite_get_rc_adapter_names}},
  \code{\link{rsuite_get_repo_adapter_names}},
  \code{\link{rsuite_register_rc_adapter}},
  \code{\link{rsuite_register_repo_adapter}},
  \code{\link{rsuite_unregister_rc_adapter}},
  \code{\link{rsuite_update}}, \code{\link{tmpl_register}}
}
