% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_api_rsuite.R
\name{rsuite_get_os_info}
\alias{rsuite_get_os_info}
\title{Retrieves information on current OS.}
\usage{
rsuite_get_os_info()
}
\value{
named list with following items
\describe{
  \item{type}{One of windows, macos, unix. (type: character)}
  \item{platform}{One of Windows, MacOS, SunOS, RedHat, Debian. (type: character(1))}
  \item{release}{One of Solaris, MacOS, Ubuntu, Debian, Fedora, CentOS or RedHat or NA. (type: character(1))}
  \item{distrib}{Distribution release e.g. for Debian: squeeze, wheezy, jessie. (type: character(1))}
  \item{version}{Version number of the distribution. (type: character(1))}
}
}
\description{
Retrieves information on current OS.
}
\examples{
rsuite_get_os_info()

}
\seealso{
Other miscellaneous: \code{\link{rsuite_check_version}},
  \code{\link{rsuite_getLogger}},
  \code{\link{rsuite_get_rc_adapter_names}},
  \code{\link{rsuite_get_repo_adapter_names}},
  \code{\link{rsuite_register_rc_adapter}},
  \code{\link{rsuite_register_repo_adapter}},
  \code{\link{rsuite_unregister_rc_adapter}},
  \code{\link{rsuite_unregister_repo_adapter}},
  \code{\link{rsuite_update}}, \code{\link{tmpl_register}}
}
