% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_api_rsuite.R
\name{rsuite_get_rc_adapter_names}
\alias{rsuite_get_rc_adapter_names}
\title{Gets all names of known RC (revision control) adapters.}
\usage{
rsuite_get_rc_adapter_names()
}
\value{
names of registered rc adapters as character vector.
}
\description{
Gets all names of known RC (revision control) adapters.
}
\examples{
rsuite_get_rc_adapter_names()

}
\seealso{
Other miscellaneous: \code{\link{rsuite_check_version}},
  \code{\link{rsuite_getLogger}},
  \code{\link{rsuite_get_ci_adapter_names}},
  \code{\link{rsuite_get_os_info}},
  \code{\link{rsuite_get_repo_adapter_names}},
  \code{\link{rsuite_register_ci_adapter}},
  \code{\link{rsuite_register_rc_adapter}},
  \code{\link{rsuite_register_repo_adapter}},
  \code{\link{rsuite_unregister_ci_adapter}},
  \code{\link{rsuite_unregister_rc_adapter}},
  \code{\link{rsuite_unregister_repo_adapter}},
  \code{\link{rsuite_update}}, \code{\link{tmpl_register}}
}
\concept{miscellaneous}
