% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_logger.R
\name{rsuite_getLogger}
\alias{rsuite_getLogger}
\title{Retrieves RSuite logger.}
\usage{
rsuite_getLogger()
}
\value{
logger object
}
\description{
Retrieves RSuite logger.
}
\examples{
logging::loginfo("This is an INFO from RSuite", logger = rsuite_getLogger())

}
\seealso{
Other miscellaneous: \code{\link{rsuite_check_version}},
  \code{\link{rsuite_get_ci_adapter_names}},
  \code{\link{rsuite_get_os_info}},
  \code{\link{rsuite_get_rc_adapter_names}},
  \code{\link{rsuite_get_repo_adapter_names}},
  \code{\link{rsuite_register_ci_adapter}},
  \code{\link{rsuite_register_rc_adapter}},
  \code{\link{rsuite_register_repo_adapter}},
  \code{\link{rsuite_unregister_ci_adapter}},
  \code{\link{rsuite_unregister_rc_adapter}},
  \code{\link{rsuite_unregister_repo_adapter}},
  \code{\link{rsuite_update}}, \code{\link{tmpl_register}}
}
\concept{miscellaneous}
