% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_api_rsuite.R
\name{rsuite_register_repo_adapter}
\alias{rsuite_register_repo_adapter}
\title{Registers repository adapter to use for projects.}
\usage{
rsuite_register_repo_adapter(repo_adapter)
}
\arguments{
\item{repo_adapter}{object complying rsuite_repo_adapter signature.}
}
\description{
Registers repository adapter to use for projects.
}
\examples{
\donttest{
  repo_adapter <- repo_adapter_create_base("Own") # create your custom adapter
  class(repo_adapter) <- c("repo_adapter_own", class(repo_adapter))
  rsuite_register_repo_adapter(repo_adapter)
}

}
\seealso{
Other miscellaneous: \code{\link{rsuite_check_version}},
  \code{\link{rsuite_getLogger}},
  \code{\link{rsuite_get_rc_adapter_names}},
  \code{\link{rsuite_get_repo_adapter_names}},
  \code{\link{rsuite_register_rc_adapter}},
  \code{\link{rsuite_unregister_rc_adapter}},
  \code{\link{rsuite_update}}, \code{\link{tmpl_register}}
}
