% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_api_rsuite.R
\name{rsuite_unregister_ci_adapter}
\alias{rsuite_unregister_ci_adapter}
\title{Unregisters CI (continuous integration) adapter.}
\usage{
rsuite_unregister_ci_adapter(name)
}
\arguments{
\item{name}{CI adapter name to unregister.}
}
\description{
Unregisters CI (continuous integration) adapter.
}
\examples{
ci_adapter <- ci_adapter_create_base("Own") # create your custom adapter
class(ci_adapter) <- c("ci_adapter_own", class(ci_adapter))

# register it
rsuite_register_ci_adapter(ci_adapter)

# unregister it
rsuite_unregister_ci_adapter("Own")

}
\seealso{
Other miscellaneous: \code{\link{rsuite_check_version}},
  \code{\link{rsuite_getLogger}},
  \code{\link{rsuite_get_ci_adapter_names}},
  \code{\link{rsuite_get_os_info}},
  \code{\link{rsuite_get_rc_adapter_names}},
  \code{\link{rsuite_get_repo_adapter_names}},
  \code{\link{rsuite_register_ci_adapter}},
  \code{\link{rsuite_register_rc_adapter}},
  \code{\link{rsuite_register_repo_adapter}},
  \code{\link{rsuite_unregister_rc_adapter}},
  \code{\link{rsuite_unregister_repo_adapter}},
  \code{\link{rsuite_update}}, \code{\link{tmpl_register}}
}
\concept{miscellaneous}
