% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_api_rsuite.R
\name{rsuite_unregister_rc_adapter}
\alias{rsuite_unregister_rc_adapter}
\title{Unregisters RC (revision control) adapter.}
\usage{
rsuite_unregister_rc_adapter(name)
}
\arguments{
\item{name}{RC adapter name to unregister.}
}
\description{
Unregisters RC (revision control) adapter.
}
\examples{
rc_adapter <- rc_adapter_create_base("Own") # create your custom adapter
class(rc_adapter) <- c("rc_adapter_own", class(rc_adapter))

# register it
rsuite_register_rc_adapter(rc_adapter)

# unregister it
rsuite_unregister_rc_adapter("Own")

}
\seealso{
Other miscellaneous: \code{\link{rsuite_check_version}},
  \code{\link{rsuite_getLogger}},
  \code{\link{rsuite_get_rc_adapter_names}},
  \code{\link{rsuite_get_repo_adapter_names}},
  \code{\link{rsuite_register_rc_adapter}},
  \code{\link{rsuite_register_repo_adapter}},
  \code{\link{rsuite_update}}, \code{\link{tmpl_register}}
}
