% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_api_rsuite.R
\name{rsuite_unregister_repo_adapter}
\alias{rsuite_unregister_repo_adapter}
\title{Unegisters repository adapter.}
\usage{
rsuite_unregister_repo_adapter(repo_adapter_name)
}
\arguments{
\item{repo_adapter_name}{name of the repo adapter to unregister. (type: character(1))}
}
\description{
Unegisters repository adapter.
}
\examples{
\donttest{
  repo_adapter <- repo_adapter_create_base("Own") # create your custom adapter
  class(repo_adapter) <- c("repo_adapter_own", class(repo_adapter))
  rsuite_register_repo_adapter(repo_adapter)
  rsuite_unregister_repo_adapter("Own")
}

}
\seealso{
Other miscellaneous: \code{\link{rsuite_check_version}},
  \code{\link{rsuite_getLogger}},
  \code{\link{rsuite_get_os_info}},
  \code{\link{rsuite_get_rc_adapter_names}},
  \code{\link{rsuite_get_repo_adapter_names}},
  \code{\link{rsuite_register_rc_adapter}},
  \code{\link{rsuite_register_repo_adapter}},
  \code{\link{rsuite_unregister_rc_adapter}},
  \code{\link{rsuite_update}}, \code{\link{tmpl_register}}
}
