% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_api_tmpl.R
\name{tmpl_get_registered}
\alias{tmpl_get_registered}
\title{Returns all available of project/package templates}
\usage{
tmpl_get_registered()
}
\value{
names of registered project and package templates together with their
file path
}
\description{
Returns all available of project/package templates
}
\details{
Project templates have to include a PARAMETERS file
Package templates have to include the following files: DESCRIPTION

All templates can be found in folder pointed by rsuite.user_templ_path option.
}
\examples{
tmpl_get_registered()

}
\seealso{
Other in templates management: \code{\link{tmpl_start_pkg}},
  \code{\link{tmpl_start_prj}}
}
