% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_api_tmpl.R
\name{tmpl_list_registered}
\alias{tmpl_list_registered}
\title{Returns all available project/package templates}
\usage{
tmpl_list_registered()
}
\value{
names of the registered project and package templates together with their
file path
}
\description{
Returns all available project/package templates
}
\details{
Project templates have to include a PARAMETERS file
Package templates have to include the following files: DESCRIPTION

All templates can be found in folder pointed by rsuite.user_templ_path option.
}
\examples{
tmpl_list_registered()

}
\seealso{
Other in templates management: \code{\link{tmpl_start}}
}
\concept{in templates management}
